Write a file using ocurrent fs plugin
Writing to a file in the pipeline without the plugin
is a _bad_ idea. The pipeline detects the state to be changing
when you write to a file and retriggers it multiple times (which
leads to multiple slack posts in this case). Hence, use the
fs plugin which introduces an explicit "save" step in the pipeline,
and doesn't retrigger the posting.